Fixed all warnings and other small fixes

+++ sipsl-code/src/SIPUTIL.cpp
@@ -94,9 +94,9 @@ void SIPUTIL::genASideReplyFromBReply(MESSAGE* _gtor, MESSAGE* __gtor, MESSAGE* 
 		DEBOUT("_gted->getHeadTo().getC_AttUriParms()).getContent()",_gted->getHeadTo()->getC_AttUriParms().getContent())
 		char totmp[512];
 		if (__gtor->getHeadTo()->getNameUri().length()!= 0){
-			sprintf(totmp, "%s %s;tag=%x",__gtor->getHeadTo()->getNameUri().c_str(), __gtor->getHeadTo()->getC_AttSipUri().getContent().c_str(),__gtor);
+			sprintf(totmp, "%s %s;tag=%x",__gtor->getHeadTo()->getNameUri().c_str(), __gtor->getHeadTo()->getC_AttSipUri().getContent().c_str(),(unsigned int)__gtor);
 		}else {
-			sprintf(totmp, "%s;tag=%x",__gtor->getHeadTo()->getC_AttSipUri().getContent().c_str(),__gtor);
+			sprintf(totmp, "%s;tag=%x",__gtor->getHeadTo()->getC_AttSipUri().getContent().c_str(),(unsigned int)__gtor);
 		}
 		string totmpS(totmp);
 		DEBOUT("******** TO new" , totmpS)
